unify ROS+ROM into one codepath (ROS); put State in more control over Consume/Skip
