+++ src/protobuf-net/ProtoReader.ReadOnlySequence.cs
@@ -5,6 +5,7 @@ using System.Buffers;
 using System.Buffers.Binary;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 using System.Text;
 
 namespace ProtoBuf
@@ -20,20 +21,152 @@ namespace ProtoBuf
         /// <param name="context">Additional context about this serialization operation</param>
         public static ProtoReader Create(out State state, ReadOnlySequence<byte> source, TypeModel model, SerializationContext context = null)
         {
-            // disabled for performance tuning
-            //if(source.IsSingleSegment)
-            //{
-            //    return Create(out state, source.First, model, context);
-            //}
-
             var reader = ReadOnlySequenceProtoReader.GetRecycled()
                 ?? new ReadOnlySequenceProtoReader();
             reader.Init(out state, source, model, context);
             return reader;
         }
 
+        /// <summary>
+        /// Creates a new reader against a multi-segment buffer
+        /// </summary>
+        /// <param name="source">The source buffer</param>
+        /// <param name="state">Reader state</param>
+        /// <param name="model">The model to use for serialization; this can be null, but this will impair the ability to deserialize sub-objects</param>
+        /// <param name="context">Additional context about this serialization operation</param>
+        public static ProtoReader Create(out State state, ReadOnlyMemory<byte> source, TypeModel model, SerializationContext context = null)
+            => Create(out state, new ReadOnlySequence<byte>(source), model, context);
+
         private sealed class ReadOnlySequenceProtoReader : ProtoReader
         {
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            internal static string ToString(ReadOnlySpan<byte> span, int offset, int bytes)
+            {
+#if PLAT_SPAN_OVERLOADS
+                return UTF8.GetString(span.Slice(offset, bytes));
+#else
+                unsafe
+                {
+                    fixed (byte* sPtr = &MemoryMarshal.GetReference(span))
+                    {
+                        var bPtr = sPtr + offset;
+                        int chars = UTF8.GetCharCount(bPtr, bytes);
+                        string s = new string('\0', chars);
+                        fixed (char* cPtr = s)
+                        {
+                            UTF8.GetChars(bPtr, bytes, cPtr, chars);
+                        }
+                        return s;
+                    }
+                }
+#endif
+            }
+            internal static int TryParseUInt32Varint(ProtoReader @this, int offset, bool trimNegative, out uint value, ReadOnlySpan<byte> span)
+            {
+                if ((uint)offset >= (uint)span.Length)
+                {
+                    value = 0;
+                    return 0;
+                }
+
+                value = span[offset++];
+                if ((value & 0x80) == 0) return 1;
+                value &= 0x7F;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                uint chunk = span[offset++];
+                value |= (chunk & 0x7F) << 7;
+                if ((chunk & 0x80) == 0) return 2;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                chunk = span[offset++];
+                value |= (chunk & 0x7F) << 14;
+                if ((chunk & 0x80) == 0) return 3;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                chunk = span[offset++];
+                value |= (chunk & 0x7F) << 21;
+                if ((chunk & 0x80) == 0) return 4;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                chunk = span[offset++];
+                value |= chunk << 28; // can only use 4 bits from this chunk
+                if ((chunk & 0xF0) == 0) return 5;
+
+                if (trimNegative // allow for -ve values
+                    && (chunk & 0xF0) == 0xF0
+                    && offset + 4 < (uint)span.Length
+                        && span[offset] == 0xFF
+                        && span[offset + 1] == 0xFF
+                        && span[offset + 2] == 0xFF
+                        && span[offset + 3] == 0xFF
+                        && span[offset + 4] == 0x01)
+                {
+                    return 10;
+                }
+
+                ThrowOverflow(@this);
+                return 0;
+            }
+            internal static int TryParseUInt64Varint(ProtoReader @this, int offset, out ulong value, ReadOnlySpan<byte> span)
+            {
+                if ((uint)offset >= (uint)span.Length)
+                {
+                    value = 0;
+                    return 0;
+                }
+                value = span[offset++];
+                if ((value & 0x80) == 0) return 1;
+                value &= 0x7F;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                ulong chunk = span[offset++];
+                value |= (chunk & 0x7F) << 7;
+                if ((chunk & 0x80) == 0) return 2;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                chunk = span[offset++];
+                value |= (chunk & 0x7F) << 14;
+                if ((chunk & 0x80) == 0) return 3;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                chunk = span[offset++];
+                value |= (chunk & 0x7F) << 21;
+                if ((chunk & 0x80) == 0) return 4;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                chunk = span[offset++];
+                value |= (chunk & 0x7F) << 28;
+                if ((chunk & 0x80) == 0) return 5;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                chunk = span[offset++];
+                value |= (chunk & 0x7F) << 35;
+                if ((chunk & 0x80) == 0) return 6;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                chunk = span[offset++];
+                value |= (chunk & 0x7F) << 42;
+                if ((chunk & 0x80) == 0) return 7;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                chunk = span[offset++];
+                value |= (chunk & 0x7F) << 49;
+                if ((chunk & 0x80) == 0) return 8;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                chunk = span[offset++];
+                value |= (chunk & 0x7F) << 56;
+                if ((chunk & 0x80) == 0) return 9;
+
+                if ((uint)offset >= (uint)span.Length) ThrowEoF(@this);
+                chunk = span[offset];
+                value |= chunk << 63; // can only use 1 bit from this chunk
+
+                if ((chunk & ~(ulong)0x01) != 0) ThrowOverflow(@this);
+                return 10;
+            }
+
             [ThreadStatic]
             private static ReadOnlySequenceProtoReader s_lastReader;
             private ReadOnlySequence<byte>.Enumerator _source;
@@ -89,7 +222,7 @@ namespace ProtoBuf
             private protected override int ImplTryReadUInt64VarintWithoutMoving(ref State state, out ulong value)
             {
                 return state.RemainingInCurrent >= 10
-                    ? ReadOnlyMemoryProtoReader.TryParseUInt64Varint(this, state.OffsetInCurrent, out value, state.Span)
+                    ? TryParseUInt64Varint(this, state.OffsetInCurrent, out value, state.Span)
                     : ViaStackAlloc(ref state, out value);
 
                 int ViaStackAlloc(ref State s, out ulong val)
@@ -118,7 +251,7 @@ namespace ProtoBuf
                     }
 
                     if (available != 10) span = span.Slice(0, available);
-                    return ReadOnlyMemoryProtoReader.TryParseUInt64Varint(this, 0, out val, span);
+                    return TryParseUInt64Varint(this, 0, out val, span);
                 }
             }
 
@@ -167,8 +300,7 @@ namespace ProtoBuf
             private protected override string ImplReadString(ref State state, int bytes)
             {
                 return state.RemainingInCurrent >= bytes
-                    ? ReadOnlyMemoryProtoReader.ToString(Consume(
-                        ref state, bytes, out var offset), offset, bytes)
+                    ? ToString(Consume(ref state, bytes, out var offset), offset, bytes)
                     : ImplReadStringMultiSegment(ref state, bytes);
             }
 
@@ -180,7 +312,7 @@ namespace ProtoBuf
                 {
                     var span = mem.Memory.Span;
                     ImplReadBytes(ref state, span);
-                    return ReadOnlyMemoryProtoReader.ToString(span, 0, bytes);
+                    return ToString(span, 0, bytes);
                 }
             }
 
@@ -206,19 +338,15 @@ namespace ProtoBuf
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             private ReadOnlySpan<byte> Consume(ref State state, int bytes)
             {
-                var span = state.Span.Slice(state.OffsetInCurrent, bytes);
-                state.Consume(bytes);
                 Advance(bytes);
-                return span;
+                return state.Consume(bytes);
             }
 
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             private ReadOnlySpan<byte> Consume(ref State state, int bytes, out int offset)
             {
-                offset = state.OffsetInCurrent;
-                state.Consume(bytes);
                 Advance(bytes);
-                return state.Span;
+                return state.Consume(bytes, out offset);
             }
 
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -228,7 +356,7 @@ namespace ProtoBuf
             private protected override int ImplTryReadUInt32VarintWithoutMoving(ref State state, Read32VarintMode mode, out uint value)
             {
                 return state.RemainingInCurrent >= 10
-                    ? ReadOnlyMemoryProtoReader.TryParseUInt32Varint(this, state.OffsetInCurrent,
+                    ? TryParseUInt32Varint(this, state.OffsetInCurrent,
                         mode == Read32VarintMode.Signed, out value, state.Span)
                     : ViaStackAlloc(ref state, mode, out value);
 
@@ -252,7 +380,7 @@ namespace ProtoBuf
                         available += take;
                     }
                     if (available != 10) span = span.Slice(0, available);
-                    return ReadOnlyMemoryProtoReader.TryParseUInt32Varint(this, 0, m == Read32VarintMode.Signed, out val, span);
+                    return TryParseUInt32Varint(this, 0, m == Read32VarintMode.Signed, out val, span);
                 }
             }
 
@@ -271,11 +399,11 @@ namespace ProtoBuf
                     }
                 }
             }
-            
+
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             private void Skip(ref State state, int bytes)
             {
-                state.Consume(bytes);
+                state.Skip(bytes);
                 Advance(bytes);
             }
